+++ lib/annotate_rb/model_annotator/foreign_key_annotation/annotation.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module ForeignKeyAnnotation
+      class Annotation
+        HEADER_TEXT = "Foreign Keys"
+
+        def initialize(foreign_keys)
+          @foreign_keys = foreign_keys
+        end
+
+        def body
+          [
+            Components::BlankLine.new,
+            Components::Header.new(HEADER_TEXT),
+            Components::BlankLine.new,
+            *@foreign_keys,
+            Components::LineBreak.new
+          ]
+        end
+
+        def to_markdown
+          body.map(&:to_markdown).join("\n")
+        end
+
+        def to_default
+          body.map(&:to_default).join("\n")
+        end
+      end
+    end
+  end
+end
